fix: circular dep and other bugsˆ
