+++ packages/round-manager/src/features/api/payoutTokens.ts
@@ -1,6 +1,7 @@
-import { ChainId, RedstoneTokenIds } from "common";
+import { ChainId } from "common";
 import { ethers } from "ethers";
 import { Address } from "wagmi";
+import { RedstoneTokenIds } from "common/src/chain-ids";
 
 export type PayoutToken = {
   name: string;
